requirements.txt
aiogram==2.9.2
python-dotenv==0.19.0
Pillow==10.2.0
